1.15.1 item.sleep refactor
