+++ src/main/java/com/lothrazar/cyclicmagic/registry/PacketRegistry.java
@@ -42,6 +42,7 @@ import com.lothrazar.cyclicmagic.item.merchant.PacketSyncVillagerToServer;
 import com.lothrazar.cyclicmagic.item.merchant.PacketVillagerTrade;
 import com.lothrazar.cyclicmagic.item.random.PacketRandomize;
 import com.lothrazar.cyclicmagic.item.scythe.PacketScythe;
+import com.lothrazar.cyclicmagic.item.sleep.PacketSleepClient;
 import com.lothrazar.cyclicmagic.item.storagesack.PacketStorageBag;
 import com.lothrazar.cyclicmagic.item.tiletransporter.PacketChestSack;
 import com.lothrazar.cyclicmagic.net.PacketItemToggle;
@@ -50,7 +51,6 @@ import com.lothrazar.cyclicmagic.net.PacketMovePlayerColumn;
 import com.lothrazar.cyclicmagic.net.PacketMovePlayerHotbar;
 import com.lothrazar.cyclicmagic.net.PacketParticleAtPosition;
 import com.lothrazar.cyclicmagic.net.PacketPlayerFalldamage;
-import com.lothrazar.cyclicmagic.net.PacketSleepClient;
 import com.lothrazar.cyclicmagic.net.PacketSound;
 import com.lothrazar.cyclicmagic.net.PacketSwapBlock;
 import com.lothrazar.cyclicmagic.net.PacketSwapPlayerHotbar;
